requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Dance==7.0.0
WTForms==3.0.1
Werkzeug==2.3.7
email-validator==2.0.0
requests-oauthlib==2.0.0
supabase==2.3.0
python-dotenv==1.0.0
postgrest-py==0.13.0
